Module Tx.In

Transaction input

type t = {

out_hash : string;

out_n : Stdint.uint32;

script : Script.t;

witness_script : Script.data list option;

sequence : Stdint.uint32;

}

val parse : ?⁠coinbase:bool -> Bitstring.bitstring -> Bitstring.bitstring * t option

parse ~coinbase:bool bdata parses an input from bdata

val parse_all : ?⁠coinbase:bool -> Bitstring.bitstring -> Bitstring.bitstring * t list option

parse_all ~coinbase:bool bdata parses a list of inputs from bdata

val serialize : t -> string

serialize inp serializes the input inp

val serialize_all : t list -> string

serialize_all inpl serializes a list of inputs inpl

val has_witness : t -> bool

has_witness inp returns true if the input has witness data