Module Bitcoinml__Block.Header

Block header parsing / serialization module

type t = {

hash : Bitcoinml.Hash.t;

version : Stdint.int32;

prev_block : Bitcoinml.Hash.t;

merkle_root : Bitcoinml.Hash.t;

time : float;

bits : string;

nonce : Stdint.uint32;

}

val parse : ?⁠hex:bool -> string -> t option

parse ~hex:bool data parses the data and returns a t option; data could be a binary data, or an hex human readable string if ~hex is true

val serialize : ?⁠hex:bool -> t -> string

serialize ~hex:bool header serializes a block header and returns a binary string or an hex human readable string if ~hex is true

val check_target : t -> bool

check_target header checks the nbits / hash target